replaced string id to int id of furniture and terain of tiles

+++ CataVOX/Assets/DDAInterop/CTile.cs
@@ -7,8 +7,8 @@ namespace Assets.DDAInterop
 {
     public struct CTile
     {
-        public IntPtr /* char* */ ter;
-        public IntPtr /* char* */ furn;
+        public int ter;
+        public int furn;
         public Vector3Int loc;
     }
 }
